+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
+#### 0.2.20 - 03/27/2017
+* FIXED: clean up logic to write a list of locked files to delete before launching the app
+* FIXED: replace the del action with delLnk to remove only shortcuts that target inside appDir
+
 #### 0.2.19 - 03/19/2017
-* TODO: add cleanup actions, e.g. remove files according to a pattern to cleanup icon on Desktop
+* ADDED: cleanup actions, e.g. remove files according to a pattern to cleanup icon on Desktop
 
 #### 0.2.18 - 03/14/2017 
 * FIXED: updater fails to cleanup in background because running app locks .exe files